+++ app/Infrastructure/Providers/EventServiceProvider.php
@@ -6,8 +6,8 @@ namespace App\Infrastructure\Providers;
 
 use App\Modules\Approval\Api\Events\EntityApproved;
 use App\Modules\Approval\Api\Events\EntityRejected;
-use App\Modules\Invoices\Application\Listeners\InvoiceApprovedListener;
-use App\Modules\Invoices\Application\Listeners\InvoiceRejectedListener;
+use App\Modules\Invoices\Infrastructure\Listeners\InvoiceApprovedListener;
+use App\Modules\Invoices\Infrastructure\Listeners\InvoiceRejectedListener;
 use Illuminate\Auth\Events\Registered;
 use Illuminate\Auth\Listeners\SendEmailVerificationNotification;
 use Illuminate\Foundation\Support\Providers\EventServiceProvider as ServiceProvider;
