+++ app/Modules/Invoices/Infrastructure/Listeners/InvoiceApprovedListener.php
@@ -1,26 +1,25 @@
 <?php declare(strict_types=1);
 
-namespace App\Modules\Invoices\Application\Listeners;
+namespace App\Modules\Invoices\Infrastructure\Listeners;
 
-use App\Modules\Approval\Api\Events\EntityRejected;
+use App\Modules\Approval\Api\Events\EntityApproved;
 use App\Modules\Invoices\Domain\Invoice;
 use App\Modules\Invoices\Domain\Repository\InvoiceRepositoryInterface;
 
-class InvoiceRejectedListener
+class InvoiceApprovedListener
 {
     public function __construct(private readonly InvoiceRepositoryInterface $invoiceRepository)
     {
     }
 
-    public function handle(EntityRejected $event): void
+    public function handle(EntityApproved $event): void
     {
         if (Invoice::class !== $event->approvalDto->entity) {
             return;
         }
 
         $invoice = $this->invoiceRepository->find($event->approvalDto->id);
-        $invoice->reject();
+        $invoice->approve();
         $this->invoiceRepository->saveStatusValue($invoice);
     }
 }
-
