+++ app/Modules/Invoices/Infrastructure/Listeners/InvoiceRejectedListener.php
@@ -1,25 +1,26 @@
 <?php declare(strict_types=1);
 
-namespace App\Modules\Invoices\Application\Listeners;
+namespace App\Modules\Invoices\Infrastructure\Listeners;
 
-use App\Modules\Approval\Api\Events\EntityApproved;
+use App\Modules\Approval\Api\Events\EntityRejected;
 use App\Modules\Invoices\Domain\Invoice;
 use App\Modules\Invoices\Domain\Repository\InvoiceRepositoryInterface;
 
-class InvoiceApprovedListener
+class InvoiceRejectedListener
 {
     public function __construct(private readonly InvoiceRepositoryInterface $invoiceRepository)
     {
     }
 
-    public function handle(EntityApproved $event): void
+    public function handle(EntityRejected $event): void
     {
         if (Invoice::class !== $event->approvalDto->entity) {
             return;
         }
 
         $invoice = $this->invoiceRepository->find($event->approvalDto->id);
-        $invoice->approve();
+        $invoice->reject();
         $this->invoiceRepository->saveStatusValue($invoice);
     }
 }
+
